+++ lib/codeowners/storage.rb
@@ -29,6 +29,42 @@ module Codeowners
       end
     end
 
+    def team_exist?(team)
+      data[:teams].find do |record|
+        record.fetch("slug") == team
+      end
+    end
+
+    def blacklisted_team?(team)
+      data[:teams].find do |record|
+        record.fetch("slug") == team &&
+          record.fetch("blacklisted")
+      end
+    end
+
+    def teams_for(user)
+      found = data[:users].find do |record|
+        record.fetch("email") == user.email ||
+          record.fetch("name") == user.name
+      end
+
+      return [] unless found
+
+      memberships = data[:memberships].find_all do |record|
+        record.fetch("user_id") == found.fetch("id")
+      end
+      memberships.map! { |hash| hash.fetch("team_id") }
+
+      return [] if memberships.empty?
+
+      teams = data[:teams].find_all do |record|
+        !record.fetch("blacklisted") &&
+          memberships.include?(record.fetch("id"))
+      end.flatten
+
+      teams
+    end
+
     private
 
     attr_reader :path