+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Codeowners
 Simple CLI to interact with GitHub CODEOWNERS
 
+## (unreleased)
+### Added
+- [Luca Guidi] Added `codeowners guess`
+
 ## v0.0.2 - 2020-06-27
 ### Added
 - [Luca Guidi] Added pattern support to `codeowners contributors`